% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rccmSummary.R
\name{aicbic}
\alias{aicbic}
\title{AIC and BIC}
\usage{
aicbic(x, omegaks)
}
\arguments{
\item{x}{List of \eqn{K} data matrices each of dimension \eqn{n_k} x \eqn{p}.}

\item{omegaks}{\eqn{p} x \eqn{p} x \eqn{K} array of \eqn{K} number of estimated subject-level precision matrices.}
}
\value{
Numeric vector of length 2 containing the AIC and BIC values.
}
\description{
This function calculates AIC and BIC values
for the random covariance clustering model (RCCM)
}
