% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_arrayProd.R
\name{arrayprod}
\alias{arrayprod}
\title{An auxiliary R function to "array" multiply an array with a vector, kindly provided by Joris Meys}
\usage{
arrayprod(x, y)
}
\arguments{
\item{x:}{a axbxc array}

\item{y:}{a vector of length c}
}
\value{
a axb matrix. The ij-th element equals sum(x[i,j,]*y)
}
\description{
An auxiliary R function to "array" multiply an array with a vector, kindly provided by Joris Meys
}
