% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_buildConfMat.R
\name{buildConfMat,character-method}
\alias{buildConfMat,character-method}
\title{buildConfMat.character}
\usage{
\S4method{buildConfMat}{character}(x, physeq)
}
\arguments{
\item{x}{a numeric matrix of confounders}

\item{physeq}{a physeq object with a sample_data slot}
}
\value{
see buidConfMat.numeric
}
\description{
buildConfMat.character
}
