% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_buildConfMat.R
\name{buildConfMat,data.frame-method}
\alias{buildConfMat,data.frame-method}
\title{buildConfMat.data.frame}
\usage{
\S4method{buildConfMat}{data.frame}(x, n)
}
\arguments{
\item{x}{a data frame of confounders}

\item{n}{the number of rows of the count matrix}
}
\value{
see buidConfMat
}
\description{
buildConfMat.data.frame
}
