% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_buildConfMat.R
\name{buildConfMat.character}
\alias{buildConfMat.character}
\title{buildConfMat.character}
\usage{
\method{buildConfMat}{character}(confounders, n, physeq, ...)
}
\arguments{
\item{confounders}{a numeric matrix of confounders}

\item{n}{the number of rows of the count matrix}

\item{physeq}{a physeq object with a sample_data slot}

\item{...}{further arguments passed on to other methods}
}
\value{
see buidConfMat.numeric
}
\description{
buildConfMat.character
}
\seealso{
\code{\link{buidConfMat.numeric}}
}
