% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_buildConfMat.R
\name{buildConfMat.data.frame}
\alias{buildConfMat.data.frame}
\title{buildConfMat.data.frame}
\usage{
\method{buildConfMat}{data.frame}(confounders, n, ...)
}
\arguments{
\item{confounders}{a data frame of confounders}

\item{n}{the number of rows of the count matrix}

\item{...}{further arguments passed on to other methods}
}
\value{
see buidConfMat.numeric
}
\description{
buildConfMat.data.frame
}
