% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_buildConfMat.R
\name{buildConfMat.numeric}
\alias{buildConfMat.numeric}
\title{buildConfMat.numeric}
\usage{
\method{buildConfMat}{numeric}(confounders, n, ...)
}
\arguments{
\item{confounders}{a numeric matrix of confounders}

\item{n}{the number of rows of the count matrix}

\item{...}{further arguments passed on to other methods}
}
\description{
buildConfMat.numeric
}
