% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_buildDesign.R
\name{buildDesign}
\alias{buildDesign}
\title{A function to build the design matrix}
\usage{
buildDesign(sampleScore, responseFun)
}
\arguments{
\item{sampleScore}{a vector of environmental scores}

\item{responseFun}{A character string, indicating the shape
of the response function

For dynamic response function estimation, the same desing matrix
as for the quadratic one is returned.
Will throw an error when an unknown repsonse function is provided}
}
\value{
A design matrix of dimension n-by-f
}
\description{
A function to build the design matrix
}
