% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_correctXMissingness.R
\name{correctXMissingness}
\alias{correctXMissingness}
\title{Replace missing entries in X by their expectation to set their
contribution to the estimating equations to zero}
\usage{
correctXMissingness(X, mu, allowMissingness)
}
\arguments{
\item{X}{the matrix of counts}

\item{mu}{the matrix of expectations}
}
\value{
The matrix X with the NA entries replaced by the
corresponding entries in mu
}
\description{
Replace missing entries in X by their expectation to set their
contribution to the estimating equations to zero
}
\note{
This may seem like a hacky approach, but it avoids having to deal
 with NAs in functions like crossprod().
}
