% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_dLR_nb.R
\name{dLR_nb}
\alias{dLR_nb}
\title{A function that returns the value of the partial derivative of the log-likelihood ratio to alpha, keeping the response functions fixed}
\usage{
dLR_nb(Alpha, X, CC, responseFun = c("linear", "quadratic", "nonparametric",
  "dynamic"), psi, NB_params, NB_params_noLab, d, alphaK, k, centMat, nLambda,
  nLambda1s, thetaMat, muMarg, ncols, envGradEst, ...)
}
\arguments{
\item{X}{the n-by-p count matrix}

\item{CC}{a n-by-d covariate vector}

\item{responseFun}{a character string indicating the type of response function}

\item{psi}{a scalar, an importance parameter}

\item{NB_params}{Starting values for the NB_params}

\item{NB_params_noLab}{Starting values for the NB_params without label}

\item{d}{an integer, the number of covariate parameters}

\item{alphaK}{a matrix of environmental gradients of lower dimensions}

\item{k}{an integer, the current dimension}

\item{centMat}{a nLambda1s-by-d centering matrix}

\item{nLambda1s}{an integer, number of centering restrictions}

\item{thetaMat}{a matrix of size n-by-p with estimated dispersion parameters}

\item{muMarg}{an n-by-p offset matrix}

\item{ncols}{a scalar, the number of columns of X}

\item{envGradEst}{a character string, indicating how the environmental gradient should be fitted. "LR" using the likelihood-ratio criterion, or "ML" a full maximum likelihood solution}

\item{Alpha:}{a vector of length d + k*(2+(k-1)/2), the environmental gradient plus the lagrangian multipliers}

\item{nLambdas}{an integer, number of lagrangian multipliers}
}
\value{
: The value of the lagrangian and the constraining equations
}
\description{
A function that returns the value of the partial derivative of the log-likelihood ratio to alpha, keeping the response functions fixed
}
