% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_dNBabunds.R
\name{dNBabundsOld}
\alias{dNBabundsOld}
\title{A score function for the column components of the independence model
(mean relative abundances)}
\usage{
dNBabundsOld(beta, X, reg, thetas, allowMissingness, naId)
}
\arguments{
\item{beta}{a vector of length p with current abundance estimates}

\item{X}{a n-by-p count matrix}

\item{reg}{a vector of length n with library sizes estimates}

\item{thetas}{a n-by-p matrix with overdispersion estimates in the rows}

\item{allowMissingness}{A boolean, are missing values present}

\item{naId}{The numeric index of the missing values in X}
}
\value{
a vector of length p with evaluations of the score function
}
\description{
A score function for the column components of the independence model
(mean relative abundances)
}
