% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_dNBlibSizes.R
\name{dNBlibSizes}
\alias{dNBlibSizes}
\title{A score function for the row components of the independence model
(library sizes)}
\usage{
dNBlibSizes(beta, X, reg, thetas)
}
\arguments{
\item{beta}{a vector of length n with current library size estimates}

\item{X}{a n-by-p count matrix}

\item{reg}{a vector of length p with relative abundance estimates}

\item{thetas}{a n-by-p matrix with overdispersion estimates in the rows}
}
\value{
a vector of length n with evaluations of the score function
}
\description{
A score function for the row components of the independence model
(library sizes)
}
