% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_dNBllcol_constr.R
\name{dNBllcol_constr}
\alias{dNBllcol_constr}
\title{The score function of the response function for 1 taxon at the time}
\usage{
dNBllcol_constr(betas, X, reg, theta, muMarg, psi)
}
\arguments{
\item{betas}{a vector of v parameters of the
response function of a single taxon}

\item{X}{the count vector of length n}

\item{reg}{a n-by-v matrix of regressors}

\item{theta}{The dispersion parameter of this taxon}

\item{muMarg}{offset of length n}

\item{psi}{a scalar, the importance parameter

 Even though this approach does not imply normalization over the parameters
 of all taxa, it is very fast and they can be normalized afterwards}
}
\value{
A vector of length v with the evaluation of the score functions
}
\description{
The score function of the response function for 1 taxon at the time
}
