% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_dNBllcol_constr_noLab.R
\name{dNBllcol_constr_noLab}
\alias{dNBllcol_constr_noLab}
\title{The score function of the general response function}
\usage{
dNBllcol_constr_noLab(betas, X, reg, thetasMat, muMarg, psi, ...)
}
\arguments{
\item{betas}{a vector of regression parameters with length v}

\item{X}{the nxp data matrix}

\item{reg}{a matrix of regressors of dimension nxv}

\item{thetasMat}{A matrix of dispersion parameters}

\item{muMarg}{offset matrix of dimension nxp}

\item{psi}{a scalar, the importance parameter}

\item{...}{further arguments passed on to the jacobian}
}
\value{
The evaluation of the score functions (a vector length v)
}
\description{
The score function of the general response function
}
