% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_dNBllcolNP.R
\name{dNBllcolNP}
\alias{dNBllcolNP}
\title{A score function for the estimation of the parameters of a third degree GLM in a constrained RC(M) model, if the GAM fit fails}
\usage{
dNBllcolNP(beta, X, reg, theta, muMarg, ...)
}
\arguments{
\item{beta}{A vector of any length}

\item{X}{the data vector of length n}

\item{reg}{a nxlength(beta) regressor matrix}

\item{theta}{a scalar, the overdispersion}

\item{muMarg}{the offset of length n}

\item{...}{further arguments passed on to the jacobian}
}
\value{
A vector of the same length as beta with evaluations of the score function
}
\description{
A score function for the estimation of the parameters of a third degree GLM in a constrained RC(M) model, if the GAM fit fails
}
