% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.discrete}
\alias{.discrete}
\title{Extract unique elements from a given column in alphabetical order.}
\usage{
.discrete(df_data, v_col)
}
\description{
Extract unique elements from a given column in alphabetical order.
}
\keyword{internal}
