% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.drop_class}
\alias{.drop_class}
\title{Drop classes matching a regular expression from an object.}
\usage{
.drop_class(obj_any, chr_pattern)
}
\arguments{
\item{obj_any}{Any R object.}

\item{chr_pattern}{A regular expression matching the classes to be dropped.}
}
\value{
The modified class list.
}
\description{
Drop classes matching a regular expression from an object.
}
\keyword{internal}
