% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.expand_checkbox_fields}
\alias{.expand_checkbox_fields}
\alias{.expand_checkbox_fields.rcm_metadata}
\alias{.expand_checkbox_fields.rcm_data}
\title{Expand checkbox field name to logical column names present in exported data.}
\usage{
.expand_checkbox_fields(...)

\method{.expand_checkbox_fields}{rcm_metadata}(df_metadata, vchr_fields)

\method{.expand_checkbox_fields}{rcm_data}(df_data, vchr_fields)
}
\description{
Expand checkbox field name to logical column names present in exported data.
}
\keyword{internal}
