% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-accessors.R
\name{.extract_choices}
\alias{.extract_choices}
\title{Extract choice-code mappings from radio/dropdown fields.}
\usage{
.extract_choices(chr_choices)
}
\arguments{
\item{chr_choices}{The choices string from the REDCap data dictionary.}
}
\value{
A list of codes named by choice.
}
\description{
Expects comma separated codes and corresponding choices, with each pair
separated by a pipe character.
}
\details{
e.g. 1,A|2,B|3,C
}
\keyword{internal}
