% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.form_event_map}
\alias{.form_event_map}
\title{Safely extract the form-event map from an rcm_data object}
\usage{
.form_event_map(df_data)
}
\arguments{
\item{df_data}{A dataframe created with `rcm`.}
}
\value{
The form-event map data frame.
}
\description{
Safely extract the form-event map from an rcm_data object
}
