% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.html_to_cli}
\alias{.html_to_cli}
\title{Translate HTML for Display with the `cli` Package}
\usage{
.html_to_cli(chr_html)
}
\arguments{
\item{chr_html}{HTML code to be stripped.}
}
\value{
Character string for display with `cli::cli_text()`.
}
\description{
Strips tags and styles, replacing these with styling codes used by `cli`.
}
\details{
As part of the `rcm` package, this function intends to handle only those
HTML tags that are likely to appear in a REDCap field label.
}
\keyword{internal}
