% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.metadata}
\alias{.metadata}
\title{Safely extract metadata from an rcm_data object}
\usage{
.metadata(df_data)
}
\arguments{
\item{df_data}{A dataframe created with `rcm`.}
}
\value{
The metadata data frame.
}
\description{
Safely extract metadata from an rcm_data object
}
