% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.strip_attributes}
\alias{.strip_attributes}
\title{Drop attributes matching a regular expression from an object.}
\usage{
.strip_attributes(obj_any, chr_pattern)
}
\arguments{
\item{obj_any}{Any R object.}

\item{chr_pattern}{A regular expression matching the attributes to be dropped.}
}
\value{
The modified class list.
}
\description{
Drop attributes matching a regular expression from an object.
}
\keyword{internal}
