% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.switch_regex}
\alias{.switch_regex}
\title{Like switch() but match value against regular expressions}
\usage{
.switch_regex(chr_string, ...)
}
\description{
Like switch() but match value against regular expressions
}
\keyword{internal}
