% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_extractDisp.R
\name{extractDisp}
\alias{extractDisp}
\title{A function that returns dispersions belonging to a RCM object}
\usage{
extractDisp(rcm, E)
}
\arguments{
\item{rcm}{an RCM object}

\item{E}{a matrix of expectations}
}
\value{
a matrix of dispersions
}
\description{
A function that returns dispersions belonging to a RCM object
}
