% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_getDevMat.R
\name{getDevMat}
\alias{getDevMat}
\title{A function to calculate the matrix of deviance residuals for given mean and dispersion matrix}
\usage{
getDevMat(X, thetaMat, mu)
}
\arguments{
\item{X}{the data matrix}

\item{thetaMat}{the matrix of dispersions}

\item{mu}{the matrix of means}
}
\value{
The matrix of deviance residuals
}
\description{
A function to calculate the matrix of deviance residuals for given mean and dispersion matrix
}
