% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_getDistCoord.R
\name{getDistCoord}
\alias{getDistCoord}
\title{Get coordinates of a distance object of n observations for the provided indices}
\usage{
getDistCoord(indices, n)
}
\arguments{
\item{indices}{The row indices for which distance indices are wanted}

\item{n}{The total number of objects in the distance matrix}
}
\value{
a vector of coordinates
}
\description{
Get coordinates of a distance object of n observations for the provided indices
}
