% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_getInflCol.R
\name{getInflCol}
\alias{getInflCol}
\title{A function to extract the influence for a given parameter index}
\usage{
getInflCol(score, InvJac, taxon)
}
\arguments{
\item{score}{a score matrix}

\item{InvJac}{The inverted jacobian}

\item{taxon}{The taxon name or index}
}
\value{
A matrix with all observations' influence on the given taxon
}
\description{
A function to extract the influence for a given parameter index
}
