% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_getInt.R
\name{getInt}
\alias{getInt}
\title{Integrate the spline of an vgam object}
\usage{
getInt(coef, spline, sampleScore, stop.on.error = FALSE, ...)
}
\arguments{
\item{coef}{A vector of coefficients}

\item{spline}{The cubic smoothing spline}

\item{sampleScore}{the observed environmental scores}

\item{stop.on.error}{see ?integrate}

\item{...}{additional arguments passed on to integrate()}
}
\value{
a scalar, the value of the integral
}
\description{
Integrate the spline of an vgam object
}
