% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_getLogLik.R
\name{getLogLik}
\alias{getLogLik}
\title{A function to extract the logged likelihood of every cell
of the data matrix from an RCM object}
\usage{
getLogLik(rcm, Dim)
}
\arguments{
\item{rcm}{an RCM object}

\item{Dim}{A vector of integers indicating which dimensions to take along,
or Inf for the saturated model, or 0 for the independence model}
}
\value{
A matrix with logged likelihood of the size of the data matrix
}
\description{
A function to extract the logged likelihood of every cell
of the data matrix from an RCM object
}
