% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_getModelMat.R
\name{getModelMat}
\alias{getModelMat}
\title{A function to construct a model matrix of a certain degree}
\usage{
getModelMat(y, degree)
}
\arguments{
\item{y}{the variable}

\item{degree}{the degree}
}
\value{
A model matrix with degree+1 columns and as many rows as lenght(y)
}
\description{
A function to construct a model matrix of a certain degree
}
