% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_heq_nb_jac.R
\name{heq_nb_jac}
\alias{heq_nb_jac}
\title{The jacobian of the linear equality constraints}
\usage{
heq_nb_jac(Alpha, alphaK, d, k, centMat, ...)
}
\arguments{
\item{Alpha}{the current estimate of the environmental gradient}

\item{alphaK}{a matrix with the environmental gradients
of the lower dimensions}

\item{d}{an integer, the number of environmental variables,
including dummies}

\item{k}{an integer, the current dimension}

\item{centMat}{a centering matrix}

\item{...}{further arguments for other methods, not needed in this one}
}
\value{
The jacobian matrix
}
\description{
The jacobian of the linear equality constraints
}
