% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_indentPlot.R
\name{indentPlot}
\alias{indentPlot}
\title{Functions to indent the plot to include the entire labels}
\usage{
indentPlot(plt, xInd = 0, yInd = 0)
}
\arguments{
\item{plt}{a ggplot object}

\item{xInd}{a scalar or a vector of length 2,
specifying the indentation left and right of the plot to allow for the labels
 to be printed entirely}

\item{yInd}{a a scalar or a vector of length 2,
specifying the indentation top and bottom of the plot
 to allow for the labels to be printed entirely}
}
\value{
a ggplot object, squared
}
\description{
Functions to indent the plot to include the entire labels
}
