% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_JacCol_constr_noLab.R
\name{JacCol_constr_noLab}
\alias{JacCol_constr_noLab}
\title{The jacobian of the estimating function for the response function
 without taxon labels}
\usage{
JacCol_constr_noLab(betas, X, reg, thetasMat, muMarg, psi, n, v, preFabMat)
}
\arguments{
\item{betas}{a vector of regression parameters with length v}

\item{X}{the nxp data matrix}

\item{reg}{a matrix of regressors of dimension nxv}

\item{thetasMat}{A matrix of dispersion parameters}

\item{muMarg}{offset matrix of dimension nxp}

\item{psi}{a scalar, the importance parameter}

\item{n}{an integer, number of rows of X}

\item{v}{an integer, the number of parameters of the response function}

\item{preFabMat}{a prefabricated matrix}
}
\value{
The jacobian (a v-by-v matrix)
}
\description{
The jacobian of the estimating function for the response function
 without taxon labels
}
