% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_NBcolInfl.R
\name{NBcolInfl}
\alias{NBcolInfl}
\title{The influence function for the column scores}
\usage{
NBcolInfl(rcm, Dim = 1)
}
\arguments{
\item{rcm}{an rcm object}

\item{Dim}{the required dimension}
}
\value{
A list with components
\item{score}{a matrix with components of the score function}
\item{InvJac}{A square matrix of dimension p with the components of the
 Jacobian related to the column scores}
}
\description{
The influence function for the column scores
}
