% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_NBjacobianAbunds.R
\name{NBjacobianAbunds}
\alias{NBjacobianAbunds}
\title{A jacobian function for the column components of the independence model
(mean relative abundances)}
\usage{
NBjacobianAbunds(beta, X, reg, thetas)
}
\arguments{
\item{beta}{a vector of length p with current abundance estimates}

\item{X}{a n-by-p count matrix}

\item{reg}{a vector of length n with library sizes estimates}

\item{thetas}{a n-by-p matrix with overdispersion estimates in the rows}
}
\value{
a diagonal matrix of dimension p with evaluations
of the jacobian function
}
\description{
A jacobian function for the column components of the independence model
(mean relative abundances)
}
