% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_NBjacobianCol.R
\name{NBjacobianCol}
\alias{NBjacobianCol}
\title{A jacobian function for the estimation of the column scores in an unconstrained RC(M) model}
\usage{
NBjacobianCol(beta, X, reg, thetas, muMarg, k, n, p, colWeights, nLambda, cMatK,
  preFabMat, Jac)
}
\arguments{
\item{beta}{vector of length p+1+1+(k-1): p row scores, 1 centering, one normalization  and (k-1) orhtogonality lagrangian multipliers}

\item{X}{the nxp data matrix}

\item{reg}{a nx1 regressor matrix: outer product of rowScores and psis}

\item{thetas}{nxp matrix with the dispersion parameters (converted to matrix for numeric reasons)}

\item{muMarg}{the nxp offset}

\item{k}{an integer, the dimension of the RC solution}

\item{n}{an integer, the number of samples}

\item{p}{an integer, the number of taxa}

\item{colWeights}{the weights used for the restrictions}

\item{nLambda}{an integer, the number of restrictions}

\item{cMatK}{the lower dimensions of the colScores}

\item{preFabMat}{a prefab matrix, (1+X/thetas)}

\item{Jac}{an empty Jacobian matrix}
}
\value{
A matrix of dimension p+1+1+(k-1) with evaluations of the Jacobian
}
\description{
A jacobian function for the estimation of the column scores in an unconstrained RC(M) model
}
