% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_NBjacobianPsi.R
\name{NBjacobianPsi}
\alias{NBjacobianPsi}
\title{A jacobian function for the psi of a given dimension (importance parameters)}
\usage{
NBjacobianPsi(beta, X, reg, muMarg, theta, preFabMat)
}
\arguments{
\item{beta}{a scalar, the current estimate}

\item{X}{the n-by-p count matrix}

\item{reg}{the regressor matrix, the outer product of current row and column scores}

\item{muMarg}{the nxp offset matrix}

\item{theta}{a n-by-p matrix with the dispersion parameters}

\item{preFabMat}{a prefab matrix, (1+X/thetas)}
}
\value{
The evaluation of the jacobian function at beta, a 1-by-1 matrix
}
\description{
A jacobian function for the psi of a given dimension (importance parameters)
}
