% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_NBpsiInfl.R
\name{NBpsiInfl}
\alias{NBpsiInfl}
\title{The influence function for the psis}
\usage{
NBpsiInfl(rcm, Dim = 1)
}
\arguments{
\item{rcm}{an rcm object}

\item{Dim}{the required dimensions}
}
\value{
The influence of every single observation
 on the psi value of this dimension
}
\description{
The influence function for the psis
}
