% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_NBrowInfl.R
\name{NBrowInfl}
\alias{NBrowInfl}
\title{The influence function for the row scores}
\usage{
NBrowInfl(rcm, Dim = 1)
}
\arguments{
\item{rcm}{an rcm object}

\item{Dim}{the required dimension}
}
\value{
A list with components
\item{score}{a matrix with components of the score function}
\item{InvJac}{A square matrix of dimension n with the components of the Jacobian related to the row scores}
}
\description{
The influence function for the row scores
}
