% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{rcm_annotation}
\alias{rcm_annotation}
\alias{rcm_annotation.rcm_field}
\alias{rcm_annotation.rcm_data}
\title{Extract the REDCap Field Annotation}
\usage{
rcm_annotation(...)

\method{rcm_annotation}{rcm_field}(v)

\method{rcm_annotation}{rcm_data}(df_data, chr_field)
}
\description{
Extract the REDCap Field Annotation
}
