% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-event.R
\name{rcm_arms_by_event}
\alias{rcm_arms_by_event}
\alias{rcm_arms_by_event.rcm_form_event_map}
\title{List REDCap arms given an event.}
\usage{
rcm_arms_by_event(...)

\method{rcm_arms_by_event}{rcm_form_event_map}(df_form_event_map, chr_event)
}
\description{
List REDCap arms given an event.
}
