% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoconvert.R
\name{rcm_autoconvert}
\alias{rcm_autoconvert}
\alias{rcm_autoconvert.rcm_data}
\alias{rcm_autoconvert.rcm_field}
\alias{rcm_autoconvert.default}
\title{Safely typecast REDCap fields given types in data dictionary.}
\usage{
rcm_autoconvert(...)

\method{rcm_autoconvert}{rcm_data}(df_data)

\method{rcm_autoconvert}{rcm_field}(v_field)

\method{rcm_autoconvert}{default}(obj)
}
\description{
Safely typecast REDCap fields given types in data dictionary.
}
