% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-accessors.R
\name{rcm_events}
\alias{rcm_events}
\alias{rcm_events.rcm_field}
\alias{rcm_events.rcm_data}
\title{Extract the REDCap events at which the field's form is present}
\usage{
rcm_events(...)

\method{rcm_events}{rcm_field}(v)

\method{rcm_events}{rcm_data}(df_data, chr_field)
}
\description{
Extract the REDCap events at which the field's form is present
}
