% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.R
\name{rcm_field}
\alias{rcm_field}
\alias{rcm_field.rcm_field}
\alias{rcm_field.rcm_data}
\title{Extract the REDCap Field Name}
\usage{
rcm_field(...)

\method{rcm_field}{rcm_field}(v)

\method{rcm_field}{rcm_data}(df_data, chr_field)
}
\description{
Extract the REDCap Field Name
}
