% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-arm.R
\name{rcm_fields_by_arm}
\alias{rcm_fields_by_arm}
\alias{rcm_fields_by_arm.rcm_data}
\title{List REDCap fields given an arm.}
\usage{
rcm_fields_by_arm(...)

\method{rcm_fields_by_arm}{rcm_data}(df_data, int_arm)
}
\description{
List REDCap fields given an arm.
}
