% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-event.R
\name{rcm_fields_by_event}
\alias{rcm_fields_by_event}
\alias{rcm_fields_by_event.rcm_data}
\title{List REDCap fields given an event.}
\usage{
rcm_fields_by_event(...)

\method{rcm_fields_by_event}{rcm_data}(df_data, chr_event)
}
\description{
List REDCap fields given an event.
}
