% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-type.R
\name{rcm_fields_by_type}
\alias{rcm_fields_by_type}
\alias{rcm_fields_by_type.rcm_metadata}
\alias{rcm_fields_by_type.rcm_data}
\title{List REDCap fields given a field type.}
\usage{
rcm_fields_by_type(...)

\method{rcm_fields_by_type}{rcm_metadata}(df_metadata, chr_type)

\method{rcm_fields_by_type}{rcm_data}(df_data, chr_type)
}
\description{
List REDCap fields given a field type.
}
