% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-arm.R
\name{rcm_forms_by_arm}
\alias{rcm_forms_by_arm}
\alias{rcm_forms_by_arm.rcm_form_event_map}
\alias{rcm_forms_by_arm.rcm_data}
\title{List REDCap instruments given an arm.}
\usage{
rcm_forms_by_arm(...)

\method{rcm_forms_by_arm}{rcm_form_event_map}(df_form_event_map, int_arm)

\method{rcm_forms_by_arm}{rcm_data}(df_data, int_arm)
}
\description{
List REDCap instruments given an arm.
}
