% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-field.R
\name{rcm_forms_by_field}
\alias{rcm_forms_by_field}
\alias{rcm_forms_by_field.rcm_metadata}
\alias{rcm_forms_by_field.rcm_field}
\title{List REDCap instruments given a REDCap field.}
\usage{
rcm_forms_by_field(...)

\method{rcm_forms_by_field}{rcm_metadata}(df_metadata, chr_field)

\method{rcm_forms_by_field}{rcm_field}(rcm_field)
}
\description{
List REDCap instruments given a REDCap field.
}
