% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-accessors.R
\name{rcm_list_events}
\alias{rcm_list_events}
\alias{rcm_list_events.rcm_data}
\alias{rcm_list_events.rcm_form_event_map}
\title{Extract the List of REDCap Events}
\usage{
rcm_list_events(...)

\method{rcm_list_events}{rcm_data}(df_data)

\method{rcm_list_events}{rcm_form_event_map}(df_form_event_map)
}
\description{
Extract the List of REDCap Events
}
