% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-accessors.R
\name{rcm_list_fields}
\alias{rcm_list_fields}
\alias{rcm_list_fields.rcm_data}
\alias{rcm_list_fields.rcm_metadata}
\title{Extract the List of REDCap Fields}
\usage{
rcm_list_fields(...)

\method{rcm_list_fields}{rcm_data}(df_data)

\method{rcm_list_fields}{rcm_metadata}(df_metadata)
}
\description{
Extract the List of REDCap Fields
}
