% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-accessors.R
\name{rcm_list_forms}
\alias{rcm_list_forms}
\alias{rcm_list_forms.rcm_data}
\alias{rcm_list_forms.rcm_metadata}
\title{Extract the List of REDCap Instruments}
\usage{
rcm_list_forms(...)

\method{rcm_list_forms}{rcm_data}(df_data)

\method{rcm_list_forms}{rcm_metadata}(df_metadata)
}
\description{
Extract the List of REDCap Instruments
}
