% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-accessors.R
\name{rcm_list_repeating}
\alias{rcm_list_repeating}
\alias{rcm_list_repeating.rcm_data}
\title{Extract the List of Repeating Forms}
\usage{
rcm_list_repeating(...)

\method{rcm_list_repeating}{rcm_data}(df_data)
}
\description{
Extract the List of Repeating Forms
}
