% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-accessors.R
\name{rcm_list_types}
\alias{rcm_list_types}
\alias{rcm_list_types.rcm_data}
\alias{rcm_list_types.rcm_metadata}
\title{Extract the List of Field Types}
\usage{
rcm_list_types(...)

\method{rcm_list_types}{rcm_data}(df_data)

\method{rcm_list_types}{rcm_metadata}(df_metadata)
}
\description{
Extract the List of Field Types
}
