% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casting.R
\name{rcm_logical}
\alias{rcm_logical}
\title{Cast numerically coded yes/no and true/false questions to logical.}
\usage{
rcm_logical(int_logical)
}
\description{
Cast numerically coded yes/no and true/false questions to logical.
}
