% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-accessors.R
\name{rcm_metadata}
\alias{rcm_metadata}
\alias{rcm_metadata.rcm_data}
\title{Extract Metadata from an rcm_data Object}
\usage{
rcm_metadata(...)

\method{rcm_metadata}{rcm_data}(df_data)
}
\description{
Extract Metadata from an rcm_data Object
}
