% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-accessors.R
\name{rcm_min}
\alias{rcm_min}
\alias{rcm_min.rcm_field}
\alias{rcm_min.rcm_data}
\alias{rcm_min.rcm_metadata}
\title{Extract the REDCap Validation Minimum Value}
\usage{
rcm_min(...)

\method{rcm_min}{rcm_field}(v)

\method{rcm_min}{rcm_data}(df_data, chr_field)

\method{rcm_min}{rcm_metadata}(df_metadata, chr_field)
}
\description{
Extract the REDCap Validation Minimum Value
}
