% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{rcm_select_by_form}
\alias{rcm_select_by_form}
\alias{rcm_select_by_form.rcm_data}
\title{Select fields given a form.}
\usage{
rcm_select_by_form(df_rc, ...)

\method{rcm_select_by_form}{rcm_data}(df_rc, vchr_form, .invert = FALSE)
}
\description{
Select fields given a form.
}
