% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{rcm_select_by_type}
\alias{rcm_select_by_type}
\title{Select fields given a form.}
\usage{
rcm_select_by_type(df_rc, vchr_type, .invert = FALSE)

rcm_select_by_type(df_rc, vchr_type, .invert = FALSE)
}
\description{
Select fields given a form.
}
