% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-factor.R
\name{rcm_to_factor}
\alias{rcm_to_factor}
\alias{rcm_to_factor.rcm_data}
\alias{rcm_to_factor.rcm_field_dropdown}
\alias{rcm_to_factor.rcm_field_radio}
\title{Safely typecast REDCap fields to factor.}
\usage{
rcm_to_factor(...)

\method{rcm_to_factor}{rcm_data}(df_data, chr_field)

\method{rcm_to_factor}{rcm_field_dropdown}(v_field)

\method{rcm_to_factor}{rcm_field_radio}(v_field)
}
\description{
Safely typecast REDCap fields to factor.
}
