% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-integer.R
\name{rcm_to_integer}
\alias{rcm_to_integer}
\alias{rcm_to_integer.rcm_data}
\alias{rcm_to_integer.rcm_field_text}
\alias{rcm_to_integer.rcm_field_dropdown}
\alias{rcm_to_integer.rcm_field_radio}
\alias{rcm_to_integer.rcm_field_slider}
\title{Safely typecast REDCap fields to integer.}
\usage{
rcm_to_integer(...)

\method{rcm_to_integer}{rcm_data}(df_data, chr_field)

\method{rcm_to_integer}{rcm_field_text}(v_field)

\method{rcm_to_integer}{rcm_field_dropdown}(v_field)

\method{rcm_to_integer}{rcm_field_radio}(v_field)

\method{rcm_to_integer}{rcm_field_slider}(v_field)
}
\description{
Safely typecast REDCap fields to integer.
}
