% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-numeric.R
\name{rcm_to_numeric}
\alias{rcm_to_numeric}
\alias{rcm_to_numeric.rcm_data}
\alias{rcm_to_numeric.rcm_field_text}
\alias{rcm_to_numeric.rcm_field_slider}
\alias{rcm_to_numeric.rcm_field_calc}
\title{Safely typecast REDCap fields to numeric}
\usage{
rcm_to_numeric(...)

\method{rcm_to_numeric}{rcm_data}(df_data, chr_field)

\method{rcm_to_numeric}{rcm_field_text}(v_field)

\method{rcm_to_numeric}{rcm_field_slider}(v_field)

\method{rcm_to_numeric}{rcm_field_calc}(v_field)
}
\description{
Safely typecast REDCap fields to numeric
}
