% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-text.R
\name{rcm_to_text}
\alias{rcm_to_text}
\alias{rcm_to_text.rcm_data}
\alias{rcm_to_text.rcm_field_text}
\alias{rcm_to_text.rcm_field_notes}
\title{Safely typecast text fields.}
\usage{
rcm_to_text(...)

\method{rcm_to_text}{rcm_data}(df_data, chr_field)

\method{rcm_to_text}{rcm_field_text}(v_field)

\method{rcm_to_text}{rcm_field_notes}(v_field)
}
\description{
Safely typecast text fields.
}
