% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-time.R
\name{rcm_to_time}
\alias{rcm_to_time}
\alias{rcm_to_time.rcm_data}
\alias{rcm_to_time.rcm_field_text}
\title{Safely typecast date/time/datetime fields.}
\usage{
rcm_to_time(...)

\method{rcm_to_time}{rcm_data}(df_data, chr_field)

\method{rcm_to_time}{rcm_field_text}(chr_time, tz = "America/New_York")
}
\description{
Safely typecast date/time/datetime fields.
}
