% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-accessors.R
\name{rcm_validation}
\alias{rcm_validation}
\alias{rcm_validation.rcm_field}
\alias{rcm_validation.rcm_data}
\alias{rcm_validation.rcm_metadata}
\title{Extract the REDCap Text Validation}
\usage{
rcm_validation(...)

\method{rcm_validation}{rcm_field}(v)

\method{rcm_validation}{rcm_data}(df_data, chr_field)

\method{rcm_validation}{rcm_metadata}(df_metadata, chr_field)
}
\description{
Extract the REDCap Text Validation
}
