% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_rowMultiply.R
\name{rowMultiply}
\alias{rowMultiply}
\title{A function to efficiently row multiply a matrix and a vector}
\usage{
rowMultiply(matrix, vector)
}
\arguments{
\item{matrix}{a numeric matrix of dimension a-by-b}

\item{vector}{a numeric vector of length b

t(t(matrix)*vector) but then faster}
}
\value{
a matrix, row multplied by the vector
}
\description{
A function to efficiently row multiply a matrix and a vector
}
\details{
Memory intensive but that does not matter with given matrix sizes
}
