% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_seq_k.R
\name{seq_k}
\alias{seq_k}
\title{A small auxiliary function for the length of the lambdas}
\usage{
seq_k(y, nLambda1s = 1)
}
\arguments{
\item{y}{an integer, the current dimension}

\item{nLambda1s}{the number of centering restrictions}
}
\value{
a vector containing the ranks of the current lagrangian multipliers
}
\description{
A small auxiliary function for the length of the lambdas
}
