% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_squarePlot.R
\name{squarePlot}
\alias{squarePlot}
\title{Enforces squareness off a ggplot}
\usage{
squarePlot(plt, xInd, yInd)
}
\arguments{
\item{plt}{a ggplot object}

\item{xInd}{a scalar or a vector of length 2, specifying the indentation left and right of the plot to allow for the labels to be printed entirely}

\item{yInd}{a a scalar or a vector of length 2, specifying the indentation top and bottom of the plot to allow for the labels to be printed entirely}
}
\value{
a ggplot object, squared
}
\description{
Enforces squareness off a ggplot
}
