% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_trimOnConfounders.R
\name{trimOnConfounders}
\alias{trimOnConfounders}
\title{Trim based on confounders to avoid taxa with only zero counts}
\usage{
trimOnConfounders(confounders, X, prevCutOff, minFraction, n)
}
\arguments{
\item{confounders}{a nxt confounder matrix}

\item{X}{the nxp data matrix}

\item{prevCutOff}{a scalar between 0 and 1, the prevalence cut off}

\item{minFraction}{a scalar between 0 and 1,
each taxon's total abundance should equal at least the number of samples n
 times minFraction, otherwise it is trimmed}

\item{n}{the number of samples

Should be called prior to fitting the independence model}
}
\value{
A trimmed data matrix nxp'
}
\description{
Trim based on confounders to avoid taxa with only zero counts
}
