% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_vgam.edit.R
\name{vgam.edit}
\alias{vgam.edit}
\title{A modified varsion of the vgam() function in the VGAM pacakage, to account for the particular model structure}
\usage{
vgam.edit(formula, psi,
  family = stop("argument 'family' needs to be assigned"),
  data = list(), coefstart = NULL, control = vgam.control(criterion =
  "loglikelihood", ...), offset = NULL, method = "vgam.fit", ...)
}
\arguments{
\item{formula}{The model formula}

\item{psi}{the importance parameter}

\item{family}{the family object}

\item{data}{the data frame}

\item{coefstart}{The starting values for the coefficients}

\item{control}{a list of control arguments}

\item{offset}{the offset, depending on lower dimensions}

\item{method}{A character string, the fitting method}
}
\value{
an object of class vgam
}
\description{
A modified varsion of the vgam() function in the VGAM pacakage, to account for the particular model structure
}
