% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_vgam.fit.edit.R
\name{vgam.fit.edit}
\alias{vgam.fit.edit}
\title{A modified version of the vgam.fit() function from the VGAM pacakage}
\usage{
vgam.fit.edit(x, y, psi, w = rep_len(1, nrow(x)), mf, coefstart = NULL,
  etastart = NULL, mustart = NULL, offset = 0, family,
  control = vgam.control(), Terms, smooth.labels,
  function.name = "vgam", ...)
}
\arguments{
\item{x}{the model matrix}

\item{y}{the outcome vector}

\item{psi}{the importance parameter}

\item{w}{a vector of weights}

\item{mf}{the model frame}

\item{coefstart}{vector of initial parameter estimates}

\item{offset}{the offset, based on previous dimensions}

\item{family}{the distributional family}

\item{control}{a list of control parameters}

\item{Terms, smooth.labels, function.name}{Other, orginal arguments of the vgam.fit() function}
}
\description{
A modified version of the vgam.fit() function from the VGAM pacakage
}
