% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_data.R
\docType{data}
\name{Zeller}
\alias{Zeller}
\title{Microbiomes of colorectal cancer patients and healthy controls}
\format{A phyloseq object with an OTU-table and sample data
\describe{
  \item{otu_table}{Count data matrix of 709 taxa in 194 samples}
  \item{sample_data}{Data frame of patient covariates}
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4299606/}
}
\usage{
Zeller
}
\description{
Microbiome sequencing data of colorectal cancer patients,
patients with small adenoma and healthy controls,
 together with otehr baseline covariates
}
\keyword{datasets}
