% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_addOrthProjection.R
\name{addOrthProjection}
\alias{addOrthProjection}
\title{This function adds orthogonal projections to a given plot based on coordinates or sample/taxon/variable names}
\usage{
addOrthProjection(RCMplot, sample = NULL, species = NULL, variable = NULL,
  Dims = 1:2, addLabel = FALSE, labPos = NULL)
}
\arguments{
\item{RCMplot}{the RCMplot object}

\item{sample, species, variable}{names or approximate coordinates of sample, species or variable}

\item{Dims}{the dimensions}

\item{addLabel}{a boolean, should the r-s-psi label be added?}

\item{labPos}{the position of the label. Will be calculated if not provided}
}
\value{
a geom_segment object that draws the projection
}
\description{
This function adds orthogonal projections to a given plot based on coordinates or sample/taxon/variable names
}
