% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_buildConfMat.R
\name{buildConfMat.character}
\alias{buildConfMat.character}
\title{buildConfMat.character}
\usage{
\method{buildConfMat}{character}(confounders, physeq)
}
\arguments{
\item{confounders}{a numeric matrix of confounders}

\item{physeq}{a physeq object with a sample_data slot}
}
\value{
see buidConfMat.numeric
}
\description{
buildConfMat.character
}
