% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_constrCorresp.R
\name{constrCorresp}
\alias{constrCorresp}
\title{Constrained correspondence analysis with adapted powers}
\usage{
constrCorresp(X, Y, a, b)
}
\arguments{
\item{X}{outcome matrix}

\item{Y}{constraining matrix}

\item{a, b}{see ?RCM_NB}
}
\value{
a list with eigenvalues, aliased variables and environmentam gradients
}
\description{
Constrained correspondence analysis with adapted powers
}
\details{
the vegan version, adapted for flexible coordinates
}
