% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_dNBllcolNP.R
\name{dNBllcolNP}
\alias{dNBllcolNP}
\title{Estimation of the parameters of a third degree GLM}
\usage{
dNBllcolNP(beta, X, reg, theta, muMarg, allowMissingness, naId, ...)
}
\arguments{
\item{beta}{A vector of any length}

\item{X}{the data vector of length n}

\item{reg}{a nxlength(beta) regressor matrix}

\item{theta}{a scalar, the overdispersion}

\item{muMarg}{the offset of length n}

\item{allowMissingness}{A boolean, are missing values present}

\item{naId}{The numeric index of the missing values in X}

\item{...}{further arguments passed on to the jacobian}
}
\value{
A vector of the same length as beta with evaluations
 of the score function
}
\description{
Estimation of the parameters of a third degree GLM
}
