% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_dNBpsis.R
\name{dNBpsis}
\alias{dNBpsis}
\title{A score function for the psi of a given dimension}
\usage{
dNBpsis(beta, X, reg, theta, muMarg, allowMissingness, naId, ...)
}
\arguments{
\item{beta}{a scalar, the initial estimate}

\item{X}{the n-by-p count matrix}

\item{reg}{the regressor matrix, the outer product of current row
and column scores}

\item{theta}{a n-by-p matrix with the dispersion parameters}

\item{muMarg}{the nxp offset matrix}

\item{allowMissingness}{A boolean, are missing values present}

\item{naId}{The numeric index of the missing values in X}

\item{...}{other arguments passed on to the jacobian}
}
\value{
The evaluation of the score function at beta, a scalar
}
\description{
A score function for the psi of a given dimension
}
