% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_estNPresp.R
\name{estNPresp}
\alias{estNPresp}
\title{A function to estimate the taxon-wise response functions non-parametrically, and return their fitted values}
\usage{
estNPresp(sampleScore, muMarg, X, ncols, thetas, n, coefInit,
  coefInitOverall, dfSpline, vgamMaxit, colWeights, verbose, ...)
}
\arguments{
\item{sampleScore}{a vector of length n with environmental scores}

\item{muMarg}{the offset matrix}

\item{X}{the n-by-p data matrix}

\item{ncols}{an integer, the number of columns of X}

\item{thetas}{a vector of length p with dispersion parameters}

\item{n}{an integer, the number of samples}

\item{coefInit}{a 2-by-p matrix with current taxon-wise parameter estimates}

\item{coefInitOverall}{a vector of length 2 with current overall parameters}

\item{dfSpline}{a scalar, the degrees of freedom for the smoothing spline.}

\item{vgamMaxit}{Maximal number of iterations in the fitting of the GAM model}

\item{colWeights}{a vector of length p with column weights}

\item{verbose}{a boolean, should number of failed fits be reported}

\item{...}{further arguments, passed on to the VGAM:::vgam() function

The negative binomial likelihood is still maximized, but now the response function is a non-parametric one. To avoid a perfect fit and overly flexible functions, we enforce smoothness restrictions. In practice we use a generalized additive model (GAM), i.e. with splines.  The same fitting procedure is carried out ignoring species labels. We do not normalize the parameters related to the splines: the psis can be calculated afterwards.}
}
\value{
A list with components
\item{taxonWiseFitted}{A n-by-p matrix of fitted valuesof the response curves per taxon at the observed values of the environmental scores}
\item{taxonCoef}{The fitted coefficients of the sample-wise response curves}
\item{overallFitted}{The fitted values of all taxa combined}
\item{taxonWiseFits}{A list of length p of normalized response curves of all the taxa. This may be useful to investigate the shape of the response function through plots.}
\item{psi}{The importance parameter of the dimension}
\item{taxonWise}{The taxonwise response function}
}
\description{
A function to estimate the taxon-wise response functions non-parametrically, and return their fitted values
}
