% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_getInflRow.R
\name{getInflRow}
\alias{getInflRow}
\title{A function to extract the influence of all observations on a given row score}
\usage{
getInflRow(score, InvJac, sample)
}
\arguments{
\item{score}{the score function evaluated for every observation}

\item{InvJac}{The inverse jacobian}

\item{sample}{the row score or  sample index}
}
\value{
A matrix with all observations' influence on the row score
}
\description{
A function to extract the influence of all observations on a given row score
}
