% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_getInt.R
\name{getInt}
\alias{getInt}
\title{A function to integrate the spline of an vgam object between the boundaries of observed environmental scores.}
\usage{
getInt(fitObj, sampleScore, class, stop.on.error = FALSE, ...)
}
\arguments{
\item{fitObj}{a fitted object, either gam or a vector of parameters}

\item{sampleScore}{the observed environmental scores}

\item{class}{the class of the fit}

\item{stop.on.error}{see ?integrate}

\item{...}{additional arguments passed on to integrate()}
}
\value{
a scalar, the value of the integral
}
\description{
A function to integrate the spline of an vgam object between the boundaries of observed environmental scores.
}
