% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_JacCol_constr.R
\name{JacCol_constr}
\alias{JacCol_constr}
\title{Returns the jacobian of this function}
\usage{
JacCol_constr(betas, X, reg, theta, muMarg, psi)
}
\arguments{
\item{betas}{a vector of v parameters of the response function of a single taxon}

\item{X}{the count vector of length n}

\item{reg}{a n-by-v matrix of regressors}

\item{theta}{The dispersion parameter of this taxon}

\item{muMarg}{offset of length n}

\item{psi}{a scalar, the importance parameter

Even though this approach does not imply normalization over the parameters of all taxa, it is very fast and they can be normalized afterwards}
}
\value{
The jacobian, a square symmetric matrix of dimension v
}
\description{
Returns the jacobian of this function
}
