% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_liks.R
\name{liks}
\alias{liks}
\title{Calculate the log-likelihoods of the indepence and saturated models and all fitted submodels}
\usage{
liks(rcm, Sum = FALSE)
}
\arguments{
\item{rcm}{an object of the RCM class}

\item{Sum}{a boolean, should likelihoods be summed?

Dispersions are re-estimated for every dimension of the model.}
}
\value{
a named array log-likelihoods of the independence model and all models with dimension 1 to k, or a vector with summed log-likelihoods
}
\description{
Calculate the log-likelihoods of the indepence and saturated models and all fitted submodels
}
