% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_NBalphaInfl.R
\name{NBalphaInfl}
\alias{NBalphaInfl}
\title{A function that the components of the influence functions of the environmental gradient.}
\usage{
NBalphaInfl(rcm, Dim)
}
\arguments{
\item{rcm}{an rcm object}

\item{Dim}{the required dimension}
}
\value{
An n-by-p-by-d array with the influence of every observation on every alpha parameter
}
\description{
A function that the components of the influence functions of the environmental gradient.
}
