% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_NBjacobianColNP.R
\name{NBjacobianColNP}
\alias{NBjacobianColNP}
\title{A jacobian function for the estimation f the parameters of
a third degree GLM in a constrained RC(M) model, if the GAM fit fails}
\usage{
NBjacobianColNP(beta, X, reg, theta, muMarg)
}
\arguments{
\item{beta}{vector of any length}

\item{X}{the data vector of length n}

\item{reg}{a nxlength(beta) regressor matrix}

\item{theta}{a scalar, the overdispersion}

\item{muMarg}{the offset of length n}
}
\value{
A matrix of dimension 8-by-8
}
\description{
A jacobian function for the estimation f the parameters of
a third degree GLM in a constrained RC(M) model, if the GAM fit fails
}
