% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_NBjacobianLibSizes.R
\name{NBjacobianLibSizes}
\alias{NBjacobianLibSizes}
\title{Jacobian for the raw components of the independence model}
\usage{
NBjacobianLibSizes(beta, X, reg, thetas)
}
\arguments{
\item{beta}{a vector of length n with current library size estimates}

\item{X}{a n-by-p count matrix}

\item{reg}{a vector of length p with relative abundance estimates}

\item{thetas}{a n-by-p matrix with overdispersion estimates in the rows}
}
\value{
a diagonal matrix of dimension n: the Fisher information matrix
}
\description{
Jacobian for the raw components of the independence model
}
