% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_plot.RCM.R
\name{plot.RCM}
\alias{plot.RCM}
\title{A function to plot RC(M) ordination result with the help of ggplot2}
\usage{
\method{plot}{RCM}(x, ..., Dim = c(1, 2), samColour = NULL, colLegend = if
  (Influence) paste0("Influence on\\n", samColour,
  "\\nparameter \\nin dimension", inflDim) else samColour, samShape = NULL,
  shapeLegend = samShape, samSize = 1.5, taxNum = if (all(plotType ==
  "species") || !is.null(taxRegExp)) {     ncol(x$X) } else {     10 },
  scalingFactor = NULL, plotType = c("samples", "species", "variables"),
  quadDrop = 0.995, nPoints = 1000, plotEllipse = TRUE, taxaScale = 0.5,
  Palette = if (!all(plotType == "species")) "Set1" else "Paired",
  taxLabels = !all(plotType == "species"), taxDots = FALSE,
  taxCol = "blue", taxColSingle = "blue", nudge_y = -0.08,
  square = TRUE, xInd = if (all(plotType == "samples")) c(0, 0) else
  c(-0.75, 0.75), yInd = c(0, 0), labSize = 2, taxRegExp = NULL,
  varNum = 15, alpha = TRUE, alphaRange = c(0.2, 1), arrowSize = 0.25,
  Influence = FALSE, inflDim = 1, richSupported = c("Observed", "Chao1",
  "ACE", "Shannon", "Simpson", "InvSimpson", "Fisher"), returnCoords = FALSE,
  varExpFactor = 10, manExpFactorTaxa = 0.975, nPhyl = 10,
  phylOther = c(""), legendSize = samSize, noLegend = is.null(samColour),
  crossSize = 4, contCol = c("orange", "darkgreen"), legendLabSize = 15,
  legendTitleSize = 16, axisLabSize = 14, axisTitleSize = 16,
  plotPsi = FALSE)
}
\arguments{
\item{x}{an RCM object}

\item{...}{further arguments, passed on to aes in the the ggplot() function}

\item{Dim}{A vector two dimensions to plot, defaults to the first two}

\item{samColour}{a character string, the variable to use for the colour of the sample dots}

\item{colLegend}{a character string, the legend text for the sample colour. Defaults to the name of the colour variable}

\item{samShape}{a character string, the variable to use for the shape of the sample dots}

\item{shapeLegend}{a character string, the text to use for the shapeLegend. Defaults to the name of the shape variable}

\item{samSize}{a scalar, the size of the sample dots}

\item{taxNum}{an integer, the number of taxa to be plotted}

\item{scalingFactor}{a scalar, a user supplied scaling factor for the taxon arrows If not supplied it will be calculated to make sample and taxon plots on the same scale}

\item{plotType}{a character string: which components to plot}

\item{quadDrop}{a number between 0 and 1. At this fraction of the peak height are the ellipses of the quadratic response functions drawn}

\item{nPoints}{an integer, number of points to evaluate for the ellipse}

\item{plotEllipse}{a boolean, whether to add the ellipses}

\item{taxaScale}{a scalar, by which to scale the rectangles of the quadratic taxon plot}

\item{Palette}{the colour palette}

\item{taxLabels}{a boolean, should taxon labels be plotted}

\item{taxDots}{a boolean, should taxa be plotted as dots?}

\item{taxCol}{the taxon colour}

\item{taxColSingle}{the taxon colour if there is only one}

\item{nudge_y}{a scalar, the offet for the taxon labels}

\item{square}{A boolean, should the plot be square? This is highly preferred to honestly represent differences}

\item{xInd}{a scalar or a vector of length 2, specifying the indentation left and right of the plot to allow for the labels to be printed entirely. Defaults to 0.75 at every side}

\item{yInd}{a scalar or a vector of length 2, specifying the indentation top and bottom of the plot to allow for the labels to be printed entirely. Defaults to 0 at every side}

\item{labSize}{the size of the variable and taxon labels}

\item{taxRegExp}{a character vector indicating which taxa to plot}

\item{varNum}{an integer, number of variable arrows to draw}

\item{alpha}{a boolean, should small arrows be made transparent?}

\item{alphaRange}{The range of transparency}

\item{arrowSize}{a scalar, the size of the arrows}

\item{Influence}{a boolean, should the influence of the observation on the variable be plotted}

\item{inflDim}{an integer, the dimension for which the influence should be calculated}

\item{richSupported}{A character vector of supported richness measures}

\item{returnCoords}{a boolean, should final coordinates be returned?}

\item{varExpFactor}{a scalar, the factor by which to expand the variable coordinates}

\item{manExpFactorTaxa}{a manual expansion factor for the taxa Setting it to a high value allows you to plot the taxa around the samples}

\item{nPhyl}{an integer, number of phylogenetic levels to show}

\item{phylOther}{a character vector of phylogenetic levels to be included in the "other" group}

\item{legendSize}{a size for the coloured dots in the legend}

\item{noLegend}{a boolean indicating you do not want a legend}

\item{crossSize}{the size of the central cross}

\item{contCol}{a character vector of length two, giving the low and high values of the continuous colour scale}

\item{legendLabSize}{size of the legend labels}

\item{legendTitleSize}{size of the legend title}

\item{axisLabSize}{size of the axis labels}

\item{axisTitleSize}{size of the axis title}

\item{plotPsi}{a boolean, should the value of the importace parameter be plotted?}
}
\value{
plots a ggplot2-object to output
}
\description{
A function to plot RC(M) ordination result with the help of ggplot2
}
\examples{
data(Zeller)
require(phyloseq)
tmpPhy = prune_taxa(taxa_names(Zeller)[1:100],
prune_samples(sample_names(Zeller)[1:50], Zeller))
 #Subset for a quick fit
zellerRCM = RCM(tmpPhy, k = 2, round = TRUE)
plot(zellerRCM)
}
