% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_plotRespFun.R
\name{plotRespFun}
\alias{plotRespFun}
\title{Plot the non-parametric response functions of a prespecified set of taxa}
\usage{
plotRespFun(RCM, taxa = NULL, Dim = 1, nPoints = 1000L, labSize = 2.5,
  yLoc = NULL, Palette = "Set3", adJitter = FALSE, subdivisions = 100L,
  nTaxa = 8L, angle = 90, ...)
}
\arguments{
\item{RCM}{and RCM object}

\item{taxa}{a character vector of taxa to be plotted}

\item{Dim}{the dimension to be plotted}

\item{nPoints}{the number of points to be used to plot the lines}

\item{labSize}{the label size for the variables}

\item{yLoc}{the y-location of the variables}

\item{Palette}{which color palette to use}

\item{adJitter}{should variable names be jittered to make them more readable}

\item{subdivisions}{the number of subdivisions for the integration}

\item{nTaxa}{an integer, number of taxa to plot}

\item{angle}{angle at which variable labels should be turned}

\item{...}{Other argumens passed on to the ggplot() function}
}
\description{
Plots a number of response functions over the observed range of the environmental score. If no taxa are provided those who react most strongly to the environmental score are chosen.
}
