% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_plotRespFun.R
\name{plotRespFun}
\alias{plotRespFun}
\title{Plot the non-parametric response functions of a prespecified set of taxa}
\usage{
plotRespFun(RCM, taxa = NULL, addSamples = TRUE, samShape = NULL,
  samSize = 2, Dim = 1, nPoints = 1000L, labSize = 2.5,
  yLocVar = NULL, yLocSam = NULL, Palette = "Set3", addJitter = FALSE,
  subdivisions = 50L, nTaxa = 8L, angle = 90, legendLabSize = 15,
  legendTitleSize = 16, axisLabSize = 14, axisTitleSize = 16, ...)
}
\arguments{
\item{RCM}{and RCM object}

\item{taxa}{a character vector of taxa to be plotted}

\item{addSamples}{a boolean, should sample points be shown?}

\item{samShape}{a sample variable name or a vector of length equal to the number of samples, for the sample shape}

\item{samSize}{the size of the sample dots}

\item{Dim}{the dimension to be plotted}

\item{nPoints}{the number of points to be used to plot the lines}

\item{labSize}{the label size for the variables}

\item{yLocVar}{the y-location of the variables, recycled if necessary}

\item{yLocSam}{the y-location of the samples, recycled if necessary}

\item{Palette}{which color palette to use}

\item{addJitter}{should variable names be jittered to make them more readable}

\item{subdivisions}{the number of subdivisions for the integration}

\item{nTaxa}{an integer, number of taxa to plot}

\item{angle}{angle at which variable labels should be turned}

\item{legendLabSize}{size of the legend labels}

\item{legendTitleSize}{size of the legend title}

\item{axisLabSize}{size of the axis labels}

\item{axisTitleSize}{size of the axis title}

\item{...}{Other argumens passed on to the ggplot() function}
}
\description{
Plots a number of response functions over the observed range of the environmental score. If no taxa are provided those who react most strongly to the environmental score are chosen.
}
