% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_RCM.R
\name{RCM}
\alias{RCM}
\title{A simple wrapper function for the RC(M) object for phyloseq objects and matrices, for all posible distributions it passes all argument onto the correct function.}
\usage{
RCM(dat, k, round = FALSE, distribution = "NB", prevCutOff = 0.025,
  minFraction = 0.1, rowWeights = "uniform", colWeights = "marginal",
  covariates = NULL, confounders = NULL, prevFit = NULL, ...)
}
\arguments{
\item{dat}{an nxp count matrix or a phyloseq object with an otu_table slot}

\item{k}{an integer, the number of dimensions of the RCM solution}

\item{round}{a boolean, whether to round to nearest integer. Defaults to FALSE.}

\item{distribution}{a character string, the error distribution of the RC(M) model. Defaults to NB}

\item{minFraction}{a scalar, each taxon's total abundance should equal at leat the number of samples n times minFraction, otherwise it is trimmed. Defaults to 10\%}

\item{rowWeights, colWeights}{character strings, the weighting procedures for the normalization of row and column scores. Defaults to "uniform" and "marginal" respectively}

\item{covariates}{In case "dat" is a phyloseq object, the names of the sample variables to be used as covariates in the constrained analysis, or "all" to indicate all variables to be used. In case dat is a matrix, a nxf matrix or dataframe of covariates. Character variables will be converted to factors, with a warning. Defaults to NULL, in which case an unconstrained analysis is carried out.}

\item{confounders}{In case "dat" is a phyloseq object, the names of the sample variables to be used as confounders to be filtered out. In case dat is a matrix, a nxf matrix or dataframe of confounders Character variables will be converted to factors, with a warning. Defaults to NULL, in which case no filtering occurs.}

\item{prevFit}{An object with a previous fit, normally from a lower dimension, that should be extended.}

\item{prevCutoff}{a scalar, the prevalance cutoff for the trimming. Defaults to 2.5e-2}
}
\value{
see RCM.NB()
}
\description{
Trim on prevalence and total abundance to avoid instability of the algorithm. We cannot conlcude much anyway on lowly abundant taxa
}
