% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcm-package.R, R/rcm.R
\docType{package}
\name{rcm}
\alias{rcm}
\title{REDCap Metadata Utilities}
\usage{
rcm(df_data, df_metadata)
}
\description{
Attach metadata to a REDCap data.frame to extract information from the
database in an informed manner.

An S3 Class that attaches rows of the REDCap data dictionary to the columns
of the data itself.
}
