% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-form.R
\name{rcm_arms_by_form}
\alias{rcm_arms_by_form}
\alias{rcm_arms_by_form.rcm_form_event_map}
\alias{rcm_arms_by_form.rcm_data}
\title{List REDCap arm given an instrument name.}
\usage{
rcm_arms_by_form(...)

\method{rcm_arms_by_form}{rcm_form_event_map}(df_form_event_map, chr_form)

\method{rcm_arms_by_form}{rcm_data}(df_data, chr_form)
}
\description{
List REDCap arm given an instrument name.
}
