% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-accessors.R
\name{rcm_choices}
\alias{rcm_choices}
\alias{rcm_choices.rcm_field}
\alias{rcm_choices.rcm_data}
\title{Extract the REDCap Choices or Calculations}
\usage{
rcm_choices(...)

\method{rcm_choices}{rcm_field}(v, raw = FALSE)

\method{rcm_choices}{rcm_data}(df_data, chr_field, ...)
}
\description{
Extract the REDCap Choices or Calculations
}
