% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/by-arm.R
\name{rcm_events_by_arm}
\alias{rcm_events_by_arm}
\alias{rcm_events_by_arm.rcm_form_event_map}
\title{List REDCap events given an arm.}
\usage{
rcm_events_by_arm(...)

\method{rcm_events_by_arm}{rcm_form_event_map}(df_form_event_map, int_arm)
}
\description{
List REDCap events given an arm.
}
