% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-accessors.R
\name{rcm_form}
\alias{rcm_form}
\alias{rcm_form.rcm_field}
\alias{rcm_form.rcm_data}
\alias{rcm_form.rcm_metadata}
\title{Extract the REDCap Instrument Name}
\usage{
rcm_form(...)

\method{rcm_form}{rcm_field}(v)

\method{rcm_form}{rcm_data}(df_data, chr_field)

\method{rcm_form}{rcm_metadata}(df_metadata, chr_field)
}
\description{
Extract the REDCap Instrument Name
}
