% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field-accessors.R
\name{rcm_type}
\alias{rcm_type}
\alias{rcm_type.rcm_field}
\alias{rcm_type.rcm_data}
\alias{rcm_type.rcm_metadata}
\title{Extract the REDCap Field Type}
\usage{
rcm_type(...)

\method{rcm_type}{rcm_field}(v)

\method{rcm_type}{rcm_data}(df_data, chr_field)

\method{rcm_type}{rcm_metadata}(df_metadata, chr_field)
}
\description{
Extract the REDCap Field Type
}
