% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F_residualPlot.R
\name{residualPlot}
\alias{residualPlot}
\title{Make residual plots of the taxa with the strongest response to the environmental gradient, or with the highest run statistic}
\usage{
residualPlot(RCM, Dim = 1, whichTaxa = "response", resid = "Deviance",
  numTaxa = 9, mfrow = NULL, samColour = NULL, samShape = NULL)
}
\arguments{
\item{RCM}{an RCM object}

\item{Dim}{an integer, which dimension?}

\item{whichTaxa}{a character string or a character vector, for which taxa to plot the diagnostic plots}

\item{resid}{the type of residuals to use, either "Deviance" or "Pearson"}

\item{numTaxa}{an integer, the number of taxa to plot}

\item{mfrow}{passed on to par(). If not supplied will be calculated based on numTaxa}

\item{samColour, samShape}{Vectors or character strings denoting teh sample colour and shape respectively. If character string is provided, the variables with this name is extracted from the phyloseq object in RCM}
}
\description{
Make residual plots of the taxa with the strongest response to the environmental gradient, or with the highest run statistic
}
\details{
If whichTaxa is "run" or "response" the taxa with the highest run statistics or responses are plotted, numTax indicates the number. If whichTaxa is a character vector, these are interpreted as taxon names to plot. This function is mainly meant for linear response functions, but can be used for others too. The runs test statistic from the tseries package is used.
}
